metrics, shifts ga3 query building logic into ga3.py module. metrics, adds ga4.py module for ga4 query building logic.
